Tela de comanda e tela de cadastro de produto(beta)

+++ FishAndChips/FishAndChips/src/main/java/com/br/pi/FishAndChips/Category/CategoryRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-    Category findByIdAndName(Integer id, String name);
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+    Category findByIdAndName(Long id, String name);
 
     Category findByName(String name);
